% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wstudent-datasets.R
\docType{data}
\name{all.ratio}
\alias{all.ratio}
\title{Proportions of each gender who graduated with Latin Honors}
\format{A data frame with 14 rows and 3 variables:
\describe{
    \item{classyear}{class year}
    \item{female}{the proportions of female students}
    \item{male}{the proporitons of male students}
}}
\usage{
all.ratio
}
\description{
A dataset containing male and female proportions from 2003 to 2016
}
\keyword{datasets}
