% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanData.R
\name{checkSuffix}
\alias{checkSuffix}
\title{Check Suffix in Line}
\usage{
checkSuffix(x)
}
\arguments{
\item{x}{a text line}
}
\value{
returns TRUE if the line contains any suffix
  listed in the text file 'suffix.txt'
}
\description{
Check if the line contains any suffix listed in the text
file 'suffix.txt'
}
\keyword{internal}
