% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHonorName.R
\name{getHonorName}
\alias{getHonorName}
\title{Get Section Indicated by Latin Honors}
\usage{
getHonorName(file_name, honor)
}
\arguments{
\item{file_name}{Name of a text file in inst/extdata or a datasets in a package. 
See details for file name convention.}

\item{honor}{Type of Latin honor from c("all", "summa", "magna", "cum", "none").
Pick 'all' for graduates with 'Cum Laude and above.'
Pick 'summa' for graduates with 'Summa Cum Laude.'
Pick 'magna' for graduates with 'Magna Cum Laude.'
Pick 'cum' for graduates with 'Cum Laude.'
Pick 'none' for graduates without Latin honors.}
}
\value{
Data frame of elements with indicated Latin honor.
}
\description{
This function works with .txt file in the package. It picks the 
elements that are marked with indicated Latin honors.
}
\details{
files in the package are saved as\cr
   "t0203.txt" for class of 2003, \cr
   "t0304.txt" for class of 2004, \cr
   .\cr
   .\cr
   .\cr
   "t1516.txt" for class of 2016\cr
   
   file_name can also be datasets in package such as wstudent.three or wstudent.ten
}
