% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanData.R
\name{isIllegalSuffix}
\alias{isIllegalSuffix}
\title{Check Illegal Suffix in Line}
\usage{
isIllegalSuffix(x)
}
\arguments{
\item{x}{text line}
}
\value{
returns TRUE if the line illegally contains a suffix
}
\description{
Check if a line illegally contains any suffix: containing
  a suffix but not ',' within the same line.
}
\keyword{internal}
