% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio.R
\name{ratio}
\alias{ratio}
\title{Get Proportional Table of Male and Female Students}
\usage{
ratio(input)
}
\arguments{
\item{input}{Clean data frame with a gender column. Recommend using datasets.}
}
\value{
Proportional table of female and male students for that input
}
\description{
Get a table of proportions of female and male students who 
graduated with Latin honors in that class year
}
