% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_rep.R
\name{stat_rep}
\alias{stat_rep}
\title{Show Statistical Representations}
\usage{
stat_rep(type)
}
\arguments{
\item{type}{Type of statistical representation to be shown.
Pick "prop.sum" for five-number summary.
Pick "boxplot" for box plot.
Pick "t.testing" for hypothesis test.
Pick "timeplot" for proportions over time.
Pick "annual" for annual information.}
}
\value{
Statistical representation according to specified type
}
\description{
Present statistical representations of the proportions of female and
    male Williams students who graduated with Latin honors in each year from 2003
    to 2016. The function uses internal datasets in the package. The datasets are 
    dataframes with students' name, genders, and Latin honors they received. Using 
    provided datasets saves computation time.

    Five options are available: five-number summary, box plot, hypothesis test, 
    time plot, and annual reports.
}
