% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wstudent-datasets.R
\docType{data}
\name{wstudent.eight}
\alias{wstudent.eight}
\title{Williams class of 2008 information}
\format{An object of class \code{data.frame} with 513 rows and 4 columns.}
\usage{
wstudent.eight
}
\description{
A dataset containing first names, mid/last names, Latin honnors, and genders of 
 the class of 2008.
 
 @format A data frame with 513 rows and 4 variables:
 \describe{
     \item{firstname}{first name}
     \item{mid/lastname}{middle and last names}
     \item{honor}{Latin honor a person graduated with}
     \item{gender}{gender of a student}
 }
}
\keyword{datasets}
